+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/integrations/emi/ComputerRecipeHandler.kt
@@ -1,6 +1,6 @@
 package site.siredvin.peripheralworks.integrations.emi
 
-import dan200.computercraft.shared.computer.inventory.ComputerMenuWithoutInventory
+import dan200.computercraft.shared.computer.inventory.AbstractComputerMenu
 import dev.emi.emi.api.recipe.EmiPlayerInventory
 import dev.emi.emi.api.recipe.EmiRecipe
 import dev.emi.emi.api.recipe.handler.EmiCraftContext
@@ -9,26 +9,21 @@ import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen
 import site.siredvin.peripheralworks.networking.ClientNetworking
 import site.siredvin.peripheralworks.networking.MapBasedEventMessage
 
-class ComputerRecipeHandler : EmiRecipeHandler<ComputerMenuWithoutInventory> {
-    override fun getInventory(p0: AbstractContainerScreen<ComputerMenuWithoutInventory>): EmiPlayerInventory = EmiPlayerInventory(emptyList())
+class ComputerRecipeHandler<T : AbstractComputerMenu> : EmiRecipeHandler<T> {
+    override fun getInventory(p0: AbstractContainerScreen<T>): EmiPlayerInventory = EmiPlayerInventory(emptyList())
 
     override fun supportsRecipe(p0: EmiRecipe): Boolean = true
 
     override fun canCraft(
         p0: EmiRecipe,
-        p1: EmiCraftContext<ComputerMenuWithoutInventory>,
+        p1: EmiCraftContext<T>,
     ): Boolean = true
 
     override fun craft(
         p0: EmiRecipe,
-        p1: EmiCraftContext<ComputerMenuWithoutInventory>,
+        p1: EmiCraftContext<T>,
     ): Boolean {
-        val recipe = mutableMapOf<String, Any>()
-        recipe["category"] = p0.category.id.toString()
-        recipe["id"] = p0.id?.toString() ?: "unknown"
-        recipe["inputs"] = p0.inputs.map(CommonEntrypoint::mapIngredient)
-        recipe["outputs"] = p0.outputs.map(CommonEntrypoint::mapIngredient)
-        ClientNetworking.sendToServer(MapBasedEventMessage(p1.screenHandler, "emi_recipe_paste", recipe))
+        ClientNetworking.sendToServer(MapBasedEventMessage(p1.screenHandler, "emi_recipe_paste", CommonEntrypoint.mapRecipe(p0)))
         return true
     }
 }